added RomMemory class and deactivateMemReadSection / deactivateMemWriteSection

+++ basetypes.h
@@ -8,7 +8,7 @@
 namespace nowind
 {
 
-typedef Uint32 byte;            // should be (at least) an 8 bit type
+typedef Uint8 byte;            // should be (at least) an 8 bit type
 typedef Uint32 word;            // should be (at least) an 16 bit type
 
 // 'typedef Uint8 byte' is 'safer' but slightly slower
